[update]鍵リスト自身の暗号化/復号を追加

+++ include/dynamic_mem_c.hpp
@@ -14,7 +14,6 @@
 
 #include "macro.hpp"
 
-
 class dynamic_mem_c
 {
 private:
@@ -31,10 +30,10 @@ public:
 
     void d_new(const uint size);
     void d_free();
-    int get_size() const;
+    size_t get_size() const;
     void copy(std::string &dest, const uint size) const;
     void reset();
-    bool is_empty()const;
+    bool is_empty() const;
 };
 
 #endif
